+++ composeApp/src/desktopMain/kotlin/org/example/project/presentation/gui/cards/items/SpecialityCard.kt
@@ -13,6 +13,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Divider
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextField
 import androidx.compose.runtime.Composable
@@ -40,6 +41,7 @@ fun SpecialityCard(
     item: Speciality,
     onDelete: (Int) -> Unit,
     onUpdate: (Speciality) -> Unit,
+    accessibility:Boolean,
 ) {
     var extended by remember { mutableStateOf(false) }
     var isEditing by remember { mutableStateOf(false) }
@@ -70,17 +72,21 @@ fun SpecialityCard(
                 modifier = Modifier.fillMaxWidth(),
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                Text(text = textName)
+                Text(text = textName, style = MaterialTheme.typography.body1)
                 Spacer(modifier = Modifier.weight(1F))
-                IconButton(
-                    onClick = {
-                        isEditing = !isEditing
-                        extended = true
-                    },
-                    content = { Icon(painter = painterResource(Res.drawable.edit_note_24px),null, tint = Color.Black) })
-                IconButton(
-                    onClick = { onDelete(item.id) },
-                    content = { Icon(painter = painterResource(Res.drawable.delete_24px),null, tint = Color.Black) })
+                println("Access in card: ${accessibility}")
+                if (accessibility)
+                {
+                    IconButton(
+                        onClick = {
+                            isEditing = !isEditing
+                            extended = true
+                        },
+                        content = { Icon(painter = painterResource(Res.drawable.edit_note_24px),null, tint = Color.Black) })
+                    IconButton(
+                        onClick = { onDelete(item.id) },
+                        content = { Icon(painter = painterResource(Res.drawable.delete_24px),null, tint = Color.Black) })
+                }
                 IconButton(
                     onClick = {
                         extended = !extended
@@ -98,13 +104,13 @@ fun SpecialityCard(
             if (extended) {
                 Divider(modifier = Modifier.fillMaxWidth().padding(vertical = Utilities.paddingExternal))
                 if (!isEditing) {
-                    Text("ID: $textId")
-                    Text("Специальность: $textName")
-                    Text("Обязаности: $textDuties")
+                    Text("ID: $textId", style = MaterialTheme.typography.body2)
+                    Text("Специальность: $textName", style = MaterialTheme.typography.body2)
+                    Text("Обязаности: $textDuties", style = MaterialTheme.typography.body2)
                 } else {
-                    Text("Id: $textId")
+                    Text("Id: $textId", style = MaterialTheme.typography.body2)
                     Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text("Специальность: ")
+                        Text("Специальность: ", style = MaterialTheme.typography.body2)
                         TextField(
                             value = textName,
                             onValueChange = { newText -> textName = newText },
@@ -113,7 +119,7 @@ fun SpecialityCard(
 
                     Row(verticalAlignment = Alignment.CenterVertically) {
                         Row(verticalAlignment = Alignment.CenterVertically) {
-                            Text("Обязаности: ")
+                            Text("Обязаности: ", style = MaterialTheme.typography.body2)
                             TextField(
                                 value = textDuties,
                                 onValueChange = { newText -> textDuties = newText },
