+++ composeApp/src/desktopMain/kotlin/org/example/project/presentation/gui/cards/items/StandardCard.kt
@@ -13,6 +13,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Divider
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextField
 import androidx.compose.runtime.Composable
@@ -40,6 +41,7 @@ fun StandardCard(
     item: Standard,
     onDelete: (Int) -> Unit,
     onUpdate: (Standard) -> Unit,
+    accessibility: Boolean,
 ) {
     var extended by remember { mutableStateOf(false) }
     var isEditing by remember { mutableStateOf(false) }
@@ -68,17 +70,21 @@ fun StandardCard(
                 modifier = Modifier.fillMaxWidth(),
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                Text(text = textName)
+                Text(text = textName, style = MaterialTheme.typography.body1)
                 Spacer(modifier = Modifier.weight(1F))
-                IconButton(
-                    onClick = {
-                        isEditing = !isEditing
-                        extended = true
-                    },
-                    content = { Icon(painter = painterResource(Res.drawable.edit_note_24px),null, tint = Color.Black) })
-                IconButton(
-                    onClick = { onDelete(item.id) },
-                    content = { Icon(painter = painterResource(Res.drawable.delete_24px),null, tint = Color.Black) })
+                if (accessibility)
+                {
+                    IconButton(
+                        onClick = {
+                            isEditing = !isEditing
+                            extended = true
+                        },
+                        content = { Icon(painter = painterResource(Res.drawable.edit_note_24px),null, tint = Color.Black) })
+                    IconButton(
+                        onClick = { onDelete(item.id) },
+                        content = { Icon(painter = painterResource(Res.drawable.delete_24px),null, tint = Color.Black) })
+                }
+
                 IconButton(
                     onClick = {
                         extended = !extended
@@ -96,20 +102,20 @@ fun StandardCard(
             if (extended) {
                 Divider(modifier = Modifier.fillMaxWidth().padding(vertical = Utilities.paddingExternal))
                 if (!isEditing) {
-                    Text("ID: $textId")
-                    Text("Название: $textName")
-                    Text("Компоненты: $textComponents")
+                    Text("ID: $textId", style = MaterialTheme.typography.body2)
+                    Text("Название: $textName", style = MaterialTheme.typography.body2)
+                    Text("Компоненты: $textComponents", style = MaterialTheme.typography.body2)
                 } else {
-                    Text("Id: $textId")
+                    Text("Id: $textId", style = MaterialTheme.typography.body2)
                     Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text("Название: ")
+                        Text("Название: ", style = MaterialTheme.typography.body2)
                         TextField(
                             value = textName,
                             onValueChange = { newText -> textName = newText },
                         )
                     }
                     Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text("Компоненты: ")
+                        Text("Компоненты: ", style = MaterialTheme.typography.body2)
                         TextField(
                             value = textComponents,
                             onValueChange = { newText -> textComponents = newText },
